Prueba de la logica en el mod sh_motorcycle_frontend
